add logic and error handling for adding and deleting a student

+++ src/main/java/git/cgteate/studentmanagementsystem/student/StudentService.java
@@ -1,7 +1,13 @@
 package git.cgteate.studentmanagementsystem.student;
 
+
 import lombok.AllArgsConstructor;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+import git.cgteate.studentmanagementsystem.student.exception.BadRequestException;
+import git.cgteate.studentmanagementsystem.student.exception.StudentNotFoundException;
 
 import java.util.List;
 
@@ -16,11 +22,19 @@ public class StudentService {
     }
 
     public void addStudent(Student student) {
-        // TODO: check if email is taken
+        Boolean existsEmail = studentRepository
+                .selectExistsEmail(student.getEmail());
+        if (existsEmail) {
+            throw new BadRequestException(
+                    "Email " + student.getEmail() + " taken");
+        }
         studentRepository.save(student);
     }
     public void deleteStudent(Long studentId) {
-         // TODO: check if student exists
+        if(!studentRepository.existsById(studentId)) {
+            throw new StudentNotFoundException(
+                    "Student with id " + studentId + " does not exists");
+        }
         studentRepository.deleteById(studentId);
     }
 }
